#7.4 Protecting Video Routes

+++ controllers/videoController.js
@@ -53,7 +53,6 @@ export const videoDetail = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id).populate("creator");
-        console.log(video);
         res.render("videoDetail", { pageTitle: video.title, video });
     } catch (error) {
         res.redirect(routes.home);
@@ -66,9 +65,12 @@ export const getEditVideo = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id);
-        res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        }
     } catch (error) {
-        console.log(error);
         res.redirect(routes.home);
     }
 };
@@ -92,9 +94,12 @@ export const deleteVideo = async (req, res) => {
         params: { id },
     } = req;
     try {
-        console.log(id);
-        await Video.findOneAndRemove({ _id: id });
-        console.log(id);
+        const video = await Video.findById(id);
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({ _id: id });
+        }
     } catch (error) {
         console.log(error);
     }
